Combine census charts to 1 component
